Added octave code to repo
A new directory containing octave source code was added.
The Octave source only does 1 sweep from the top at the moment,
and will print everything to png's if asked. See README in the
octsrc folder

Also, an organizational change was made to the c code.
The makefile was modified so all the object files were
generated in their own directory rather than filling up
the base directory with crap.